+++ Electrode/src/Electrode_RadialRegion.h
@@ -195,7 +195,7 @@ public:
 
 
     //! Get all of the reaction rates and parameters from Cantera
-    void extractInfoJustBorn(std::vector<int>& justBornMultiSpecies);
+    void extractInfoJustBorn(std::vector<size_t>& justBornMultiSpecies);
 
     int calcResid(double* const resid,  const ResidEval_Type_Enum evalType) ;
