+++ Electrode/src/Electrode_SimplePhaseChangeDiffusion.h
@@ -81,7 +81,8 @@ public:
 
 
     void calcRate(double deltaT);
-    void extractInfo(std::vector<int>& justBornMultiSpecies);
+
+    void extractInfo(std::vector<size_t>& justBornMultiSpecies);
 
     virtual void updateState();
